ee/cloud/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dinobase-hosted"
version = "0.1.0"
description = "Dinobase Cloud — hosted service with OAuth, sync, and query API"
requires-python = ">=3.10"
license = "Elastic-2.0"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-jose[cryptography]>=3.3",
    "cryptography>=42.0",
    "supabase>=2.0",
    "python-dotenv>=1.0",
    "dinobase",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
dinobase-hosted = "dinobase_hosted.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dinobase_hosted"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
